+++ features/transaction/handler/format.go
@@ -3,15 +3,15 @@ package handler
 import "time"
 
 type TransactionReq struct {
-	UserId uint    `json:"user_id"`
-	Amount float64 `json:"amount"`
+	UserId uint `json:"user_id"`
+	Amount int  `json:"amount"`
 }
 
 type TransactionRes struct {
 	ID        uint      `json:"transaction_id"`
 	OrderID   string    `json:"order_id"`
 	UserId    uint      `json:"user_id"`
-	Amount    float64   `json:"amount"`
+	Amount    int       `json:"amount"`
 	Status    string    `json:"status"`
 	Token     string    `json:"token"`
 	Url       string    `json:"url"`
